QEDfields

Documentation for QEDfields.

QEDfields.AbstractBackgroundField

QEDfields.AbstractPlaneWaveField

QEDfields.AbstractPulseProfile

QEDfields.AbstractPulsedPlaneWaveField

QEDfields.Analytical

QEDfields.CosSquarePulse

QEDfields.GaussKronrodQuadrature

QEDfields.GaussLegendreQuadrature

QEDfields.GaussianPulse

QEDfields.InternalIntegralResult

QEDfields.InternalIntegrals

QEDfields.PhaseIntegralResult

QEDfields.PhaseIntegrals

QEDfields.PulsedPlaneWaveField

QEDfields._amplitude

QEDfields._envelope

QEDfields._internal_integrals

QEDfields._unsafe_gaussian_envelope

QEDfields._volkov_phase

QEDfields.amplitude

QEDfields.classical_nonlinearity_parameter

QEDfields.domain

QEDfields.envelope

QEDfields.generic_spectrum

QEDfields.integrate

QEDfields.oscillator

QEDfields.polarization

QEDfields.polarization_vector

QEDfields.pulse_length

QEDfields.pulse_profile

QEDfields.reference_momentum

QEDfields.AbstractBackgroundField — Type

Abstract base type for describing classical background fields.

Currently, only plane-wave fields are supported.

Interface funcitons:

polarization(::AbstractBackgroundField)::AbstractPolarization

QEDfields.AbstractPlaneWaveField — Type

AbstractPlaneWaveField

Abstract base type for plane-wave fields.

Interface functions:

_amplitude(::AbstractPlaneWaveField, ::AbstractDefinitePolarization, ::Real)

_domain(field::AbstractPlaneWaveField)

reference_momentum(::AbstractBackgroundField)::AbstractFourMomentum

classical_nonlinearity_parameter(::AbstractBackgroundField)::Real

polarization(::AbstractBackgroundField)::AbstractPolarization

Definition: plane-wave fields

Plane-wave fields are an idealization of electromagnetic fields from sources infinitely far away. Mathematically, a field is referred to as plane-wave, if any of the following equivalent points is fulfilled:

the field only depends on $\phi=k\cdot x = \omega t - \vec k \vec x$ (mostly minus metric),

points with $\vec k \vec x = \mathrm{const.}$ form parallel planes perpendicular to $\vec k$,

the field invariants vanish, i.e. $F^{\mu\nu}F_{\mu\nu} = 0$, $\epsilon_{\mu\nu\tau\lambda}F^{\mu\nu}F^{\tau\lambda}=0$.

QEDfields.AbstractPulseProfile — Type

AbstractPulseProfile

Abstract base type for pulse profiles.

Interface functions:

domain(::AbstractPulseProfile)::AbstractInterval

pulse_length(::AbstractPulseProfile)::Real

_envelope(::AbstractPulseProfile,phi::Real)::Real

QEDfields.AbstractPulsedPlaneWaveField — Type

Abstract base type for pulsed plane-wave fields.

Interface functions (from AbstractPlaneWaveField):

reference_momentum(::AbstractBackgroundField)::AbstractFourMomentum

classical_nonlinearity_parameter(::AbstractBackgroundField)::Real

polarization(::AbstractBackgroundField)::AbstractPolarization

Additional interface functions:

pulse_profile(::AbstractPulsedPlaneWaveField)::AbstractPulseProfile

QEDfields.Analytical — Type

Analytical()

General analytical solution of an integral.

QEDfields.CosSquarePulse — Type

CosSquarePulse(mom::M,pulse_length::T) where {M<:QEDbase.AbstractFourMomentum,T<:Real}

Concrete implementation of an AbstractPulsedPlaneWaveField for cos-square pulses.

Pulse shape

The pulse envelope of a cos-square pulse is defined as

\[g(\phi) = \cos^2(\frac{\pi\phi}{2\Delta\phi})\]

for $\phi\in (-\Delta\phi,\Delta\phi)$, where $\Delta\phi$ denotes the pulse_length, and zero otherwise.

QEDfields.GaussKronrodQuadrature — Type

GaussKronrodQuadrature()

Integration method using QuadGK.quadgk.

QEDfields.GaussLegendreQuadrature — Type

GaussLegendreQuadrature(order)

Integration method using Gauss-Legendre nodes provided by FastGaussQuadrature.jl

QEDfields.GaussianPulse — Type

GaussianPulse(mom::M,pulse_length::T) where {M<:QEDbase.AbstractFourMomentum,T<:Real}

Concrete implementation of an AbstractPulsedPlaneWaveField for Gaussian pulses.

Propagates along the direction given by the space-like components of the reference momentum vector kmu = (omega/speedoflight, kx, ky, kz), and omega = 2pispeedoflight/wavelength.

The time-like coordinate omega/speedofligth of k_mu defines the field's oscillation frequency.

In order to fulfill the vacuum dispersion relation, k_mu*k^mu=0 is required.

Pulse shape

The longitudinal envelope of a Gaussian pulse is defined as

\[g(\phi) = \exp(-\frac{\phi^2}{2\Delta\phi^2})\]

for $\phi\in (-\infty, \infty) and $\Delta\phi`$ denotes the pulse_length.

There is no envelope in the transverse directions.

QEDfields.InternalIntegralResult — Type

InternalIntegralResult

Result of computing the internal integral I(phi, beta1, beta2)

value: the computed integral value

error: error estimate (Nothing if not available)

QEDfields.InternalIntegrals — Type

InternalIntegrals(tuple_of_results)
InternalIntegrals(resuls1, result2, ... )

Stores the values of internal integrals.

For definite polarization:

\[I_1 = \int_0^x f(y)\,\mathrm{d}y\\ I_2 = \int_0^x f^2(y)\,\mathrm{d}y\]

For indefinite polarization:

\[ I_11 = \int_0^x f_x(y)\,\mathrm{d}y\\ I_12 = \int_0^x f_y(y)\,\mathrm{d}y\\ I_2 = \int_0^x f_x^2(y)+ f_y^2(y)\,\mathrm{d}y\]

QEDfields.PhaseIntegralResult — Type

PhaseIntegralResult

Result of computing the phase integral B(pnum,beta1,beta2).

value: the computed integral value

error: error estimate (Nothing if not available)

QEDfields.PhaseIntegrals — Type

PhaseIntegrals

Stores the results of phase integrals.

QEDfields.PulsedPlaneWaveField — Type

PulsedPlaneWaveField(
    pulse::AbstractPulseProfile,
    mom::AbstractFourMomentum,
    a0::Real,
    pol::AbstractPolarization
)

QEDfields._amplitude — Function

_amplitude(::AbstractPlaneWaveField, pol::AbstractDefinitePolarization, phi)

Interface function for background fields. Returns the value of the amplitude at a given point phi.

QEDfields._envelope — Function

_envelope(::AbstractPulseProfile, phi::Real)

Return value of the envelope without bound-check.

QEDfields._internal_integrals — Function

_internal_integrals(::AbstractBackgroundField)

Returns the tuple of internal integrals: (I1, I2) for linear polarization, and (I11,I12,I2) for elliptic polarization

QEDfields._unsafe_gaussian_envelope — Method

_unsafe_gaussian_envelope(phi::Real, dphi::Real)

The envelope of the Gaussian background field is defined according to the standard Gaussian distribution with dphi representing the distribution's standard deviation.

QEDfields._volkov_phase — Function

_volkov_phase(
    ::AbstractBackgroundField,
    ::IntegrationMethod,
    phi::Real,
    beta1::Real,
    beta2::Real
)

QEDfields.amplitude — Method

amplitude(field::AbstractPlaneWaveField, pol::AbstractDefinitePolarization, phi)

Returns the value of the amplitude for a given polarization direction and phase variable phi.

Safe implementation

In this function, a domain check is performed, i.e. if phi is in the domain of the field, the value of the amplitude is returned, and zero otherwise.

QEDfields.classical_nonlinearity_parameter — Function

classical_nonlinearity_parameter(::AbstractPlaneWaveField)

Interface function for plane-wave background fields. Return the classical nonlinearity parameter.

QEDfields.domain — Function

domain(::AbstractPulseProfile)

QEDfields.envelope — Method

envelope(pulsed_field::AbstractPulsedPlaneWaveField, phi::Real)

Return value of the envelope with bound-check.

QEDfields.generic_spectrum — Method

generic_spectrum(field::AbstractPulsedPlaneWaveField, pol::AbstractDefinitePolarization, pnum)

Return the generic spectrum of the given field, for the given polarization direction pol and a given photon number parameter pnum.

Convention

The generic spectrum is defined as the Fourier transform of the respective amplitude function for the given polarization direction:

\[\begin{align*} x-\mathrm{pol} &\to \int_{-\infty}^{\infty} g(\phi) \cos(\phi) \exp(il\phi)\\ y-\mathrm{pol} &\to \int_{-\infty}^{\infty} g(\phi) \sin(\phi) \exp(il\phi) \end{align*}\]

where $g(\phi)$ is the envelope and $l$ the photon number parameter.

QEDfields.integrate — Function

integrate(::AbstractNumericalIntegrationMethod,func,low,up)

Interface function for subtypes of AbstractNumericalIntegrationMethod.

QEDfields.oscillator — Function

oscillator(pol::AbstractPolarizaion, phi::Real)

Return the value of the base oscillator associated with a given polarization pol at a given point phi.

Convention

The current default implementation are

PolX() -> cos(phi)
PolY() -> sin(phi)
AllPol() -> (cos(phi), sin(phi))

QEDfields.polarization — Function

polarization(field::AbstractBackgroundField)

Interface function for general background fields. Return object which is subtype of AbstractPolarization.

QEDfields.polarization_vector — Method

polarization_vector(pol::AbstractPolarization, mom::QEDbase.AbstractFourMomentum)

Return the polarization vector for a given polarization and four-momentum mom. For a definite polarization, the respective LorentzVector is returned, where as for an indefinite polarization, a tuple of polarization vectors is returned.

Convention

In the current implementation, we use the base_state function for Photon provided by QEDcore.jl.

QEDfields.pulse_length — Function

pulse_length(::AbstractPulseProfile)

QEDfields.pulse_profile — Function

pulse_profile(field::AbstractPulsedPlaneWaveField)

Return subtype of AbstractPulseProfile.

QEDfields.reference_momentum — Function

reference_momentum(::AbstractPlaneWaveField)

Interface function for plane-wave background fields. Return a reference momentum indicating the direction the field is propagating.